# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gemini/conversations:
    get:
      operationId: listGeminiConversations
      tags: [gemini]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiConversation"
    post:
      operationId: createGeminiConversation
      tags: [gemini]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversation"
  /gemini/conversations/{id}:
    get:
      operationId: getGeminiConversation
      tags: [gemini]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
    delete:
      operationId: deleteGeminiConversation
      tags: [gemini]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
  /gemini/conversations/{id}/messages:
    get:
      operationId: listGeminiMessages
      tags: [gemini]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiMessage"
    post:
      operationId: sendGeminiMessage
      tags: [gemini]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}
  /gemini/generate-post:
    post:
      operationId: generatePost
      tags: [gemini]
      summary: Generate a Day 3 editorial post from a staged article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePostInput"
      responses:
        "200":
          description: Generated editorial post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedPost"
        "500":
          description: Generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
  /gemini/summarize:
    post:
      operationId: summarizeArticles
      tags: [gemini]
      summary: Summarize a batch of staged articles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SummarizeArticlesInput"
      responses:
        "200":
          description: SSE stream of article summaries
          content:
            text/event-stream: {}
...
components:
  schemas:
...
    Article:
      type: object
      properties:
        id:
          type: string
          example: "31f_1"
        title:
          type: string
          example: "New GST notification issued by CBIC"
        url:
          type: string
          format: uri
          example: "https://taxguru.in/..."
        date:
          type: string
          example: "May 18, 2025"
        category:
          type: string
          example: "GST"
        source:
          type: string
          example: "LiveMint"
        excerpt:
          type: string
          example: "The Central Board of Indirect Taxes..."
      required:
        - id
        - title
        - url
        - date
        - category
        - excerpt
...
    GeminiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    GeminiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    GeminiConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    GeminiMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    GeminiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/GeminiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
...
    GeminiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SummarizeArticlesInput:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/Article"
      required:
        - articles
    GeneratePostInput:
      type: object
      properties:
        article:
          $ref: "#/components/schemas/Article"
        firmInsight:
          type: string
          description: Optional custom firm perspective to inject into the prompt
        postFormat:
          type: string
          enum: [analysis, news, update]
          default: analysis
          description: Format of the post to generate
      required:
        - article
    GeneratedPost:
      type: object
      properties:
        postFormat:
          type: string
          enum: [analysis, news, update]
          default: analysis
        title:
          type: string
        summaryOfFacts:
          type: string
        keyTakeaways:
          type: array
          items:
            type: string
        firmPerspective:
          type: string
        articleCategory:
          type: string
        articleDate:
          type: string
        articleUrl:
          type: string
        hashtags:
          type: array
          items:
            type: string
        linkedInPost:
          type: string
        headline:
          type: string
        summary:
          type: string
        whyItMatters:
          type: string
        whatChanged:
          type: string
        effectiveDate:
          type: string
        appliesTo:
          type: string
        actionRequired:
          type: string
      required:
        - title
        - articleCategory
        - articleDate
        - articleUrl
...