# lib/api-zod/src/generated/types/generatePostInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Article } from './article';
import type { GeneratePostInputPostFormat } from './generatePostInputPostFormat';

export interface GeneratePostInput {
  article: Article;
  /** Optional custom firm perspective to inject into the prompt */
  firmInsight?: string;
  /** Format of the post to generate */
  postFormat?: GeneratePostInputPostFormat;
}